+++ miracle/data/delete.py
@@ -1,4 +1,7 @@
+from datetime import datetime
+
 from sqlalchemy import delete
+from sqlalchemy import select
 
 from miracle.models import User
 
@@ -7,10 +10,19 @@ def delete_data(task, user):
     # Delete user data from the database.
     exists = False
     with task.db.session() as session:
-        result = session.execute(delete(User).where(User.token == user))
-        exists = bool(result.rowcount)
+        row = session.execute(
+            select([User.id, User.created]).where(User.token == user)
+        ).fetchone()
+        if row:
+            created = row[1]
+            result = session.execute(delete(User).where(User.token == user))
+            exists = bool(result.rowcount)
     if exists:
         task.stats.increment('data.user.delete')
+        if created:
+            now = datetime.utcnow().replace(second=0, microsecond=0)
+            diff_hours = int(round((now - created).total_seconds() / 3600.0))
+            task.stats.timing('data.user.delete_hours', diff_hours)
     return exists