+++ miracle/api/tests.py
@@ -1,5 +1,10 @@
+from datetime import datetime, timedelta
 import json
 
+from pyramid.httpexceptions import HTTPForbidden
+import pytest
+
+from miracle.api.views import check_end_date
 
 CORS_HEADERS = {
     'Access-Control-Allow-Origin',
@@ -7,6 +12,18 @@ CORS_HEADERS = {
     'Access-Control-Allow-Headers',
     'Access-Control-Allow-Methods',
 }
+
+
+def test_check_end_date():
+    today = datetime.utcnow().date()
+    check_end_date(today + timedelta(days=90))
+    check_end_date(today + timedelta(days=1))
+    with pytest.raises(HTTPForbidden):
+        check_end_date(today)
+    with pytest.raises(HTTPForbidden):
+        check_end_date(today - timedelta(days=1))
+    with pytest.raises(HTTPForbidden):
+        check_end_date(today - timedelta(days=90))
 
 
 def test_delete(app, stats):