+++ miracle/__init__.py
@@ -1,1 +1,1 @@
-VERSION = '1.1.1.dev1'
+VERSION = '1.1.1'